requirements.txt
Flask==2.0.1
Flask-SocketIO==5.3.2
python-socketio==5.7.2
requests==2.28.1